Add button that resets board and changes sq size
I added a button that when clicked, removes all the children of the main
container holding the divs, and then shows a prompt that asks for the
size of each side of the new grid. This size will be the size of the
squares in the grid, not the size of the grid itself. This prompt is
inside a loop in order to reprompt in case the user enters a number
higher than 100 or lower than 1.